gateway/requirements.txt
flask
grpcio
mysql-connector-python
python-dotenv
protobuf
Flask-Cors
